fix: build qt designer custom plugin loading .so in Qt 5.15.x
This works around the code in PyQt that disables creating the
'libpyqt5.so' file for Qt Designer. It works by using the same logic
PyQt should have used normally, but doesn't because the CPython on
conda-forge is build without '--enable-shared'.

The entries appended to PyQt's 'pyproject.toml' are described in
https://www.riverbankcomputing.com//static/Docs/PyQt-builder/pyproject_toml.html

Signed-off-by: flow <flowlnlnln@gmail.com>